requirements/base.txt
# Django and Core Framework
Django==5.2.2 - https://github.com/django/django

# Task Queue
django-q2==1.8.0
redis==5.2.1

# Django REST Framework
djangorestframework==3.16.0  # https://github.com/encode/django-rest-framework
djangorestframework_simplejwt==5.5.0  # https://github.com/jazzband/djangorestframework-simplejwt

# Database
psycopg==3.2.9  # https://github.com/psycopg/psycopg

# Django Extensions
django-cors-headers==4.7.0  # https://github.com/adamchainz/django-cors-headers
django-filter==25.1  # https://github.com/carltongibson/django-filter 
django-redis==5.4.0  # https://github.com/jazzband/django-redis
drf-yasg==1.21.10  # https://github.com/axnsan12/drf-yasg

# Data and Services
redis==5.2.1  # https://github.com/redis/redis-py
PyYAML==6.0.2  # https://github.com/yaml/pyyaml
requests==2.32.4  # https://github.com/psf/requests

# Logging
python-json-logger==3.3.0  # JSON formatted logs
structlog==24.1.0  # Structured logging

# ULID
python-ulid==3.0.0

# Testing Framework
pytest==8.4.1  # https://docs.pytest.org/en/stable/
pytest-django==4.11.1  # https://github.com/pytest-dev/pytest-django
pytest-sugar==1.0.0  # https://github.com/Frozenball/pytest-sugar

# Coverage
factory-boy==3.3.3  # https://github.com/FactoryBoy/factory_boy
model-bakery==1.20.5  # https://github.com/model-bakers/model_bakery
coverage==7.9.1  # https://github.com/nedbat/coveragepy
django-coverage-plugin==3.1.0  # https://github.com/nedbat/django_coverage_plugin

# Mock
requests-mock==1.12.1  # https://github.com/jamielennox/requests-mock

# Time Mocking
freezegun=1.5.3

# vObject
# ------------------------------------------------------------------------------
vobject==0.9.6.1

# AI

# Money
py-moneyed==3.0